New artifactId and package to be able to coexist
with previous version.
