--- src/main/java/no/digipost/cache/fallback/marshall/SerializingMarshaller.java
@@ -1,37 +1,0 @@
-/**
- * Copyright (C) Posten Norge AS
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package no.digipost.cache.fallback.marshall;
-
-import java.io.*;
-
-public class SerializingMarshaller<T extends Serializable> implements Marshaller<T> {
-
-	@Override
-	public T read(InputStream input) throws Exception {
-		try (ObjectInputStream ois = new ObjectInputStream(input)) {
-			@SuppressWarnings("unchecked")
-			T unserialized = (T) ois.readObject();
-			return unserialized;
-		}
-	}
-
-	@Override
-	public void write(T toWrite, OutputStream output) throws IOException {
-		try (ObjectOutputStream oos = new ObjectOutputStream(output)) {
-			oos.writeObject(toWrite);
-		}
-	}
-}
