+++ src/main/java/no/digipost/cache2/loader/Callables.java
@@ -1,0 +1,38 @@
+/**
+ * Copyright (C) Posten Norge AS
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package no.digipost.cache2.loader;
+
+import java.util.concurrent.Callable;
+
+/**
+ * Utilities to work with {@link Callable}s as cache value loaders.
+ */
+public final class Callables {
+
+	/**
+	 * "Promote" a {@code Callable} as a {@code Loader}. The resulting
+	 * {@code Loader} ignores the argument given to {@link Loader#load(Object)}
+	 * and delegates to the {@code Callable}.
+	 */
+	public static <K, V> Loader<K, V> toLoader(final Callable<V> callable) {
+		return new Loader<K, V>() {
+			@Override
+			public V load(K key) throws Exception {
+				return callable.call();
+			}
+		};
+	}
+}
